r=int(input()) pi=3.14 area=pi*r*r print("%.2f"%area)

+++ Area_of_a_circle.py
@@ -1,4 +1,4 @@
-r=float(input())
+r=int(input())
 pi=3.14
 area=pi*r*r
 print("%.2f"%area)
